# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Monorepo Fastify Vue API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "HttpError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "additionalProperties": false
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {}
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "details"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/api/v1/me/authorization": {
      "get": {
        "tags": [
          "Authorization"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email"
                      ],
                      "additionalProperties": false
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true,
                            "maximum": 9007199254740991
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "rules": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "action": {
                            "type": "string",
                            "enum": [
                              "manage",
                              "create",
                              "read",
                              "update",
                              "delete",
                              "assign"
                            ]
                          },
                          "subject": {
                            "type": "string",
                            "enum": [
                              "all",
                              "Task",
                              "Profile",
                              "User",
                              "Role",
                              "AbilityRule",
                              "AuditLog"
                            ]
                          },
                          "fields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "conditions": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "inverted": {
                            "type": "boolean"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "action",
                          "subject"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "authorizationVersion": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "user",
                    "roles",
                    "rules",
                    "authorizationVersion"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/authorization/catalog": {
      "get": {
        "tags": [
          "Authorization"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "manage",
                          "create",
                          "read",
                          "update",
                          "delete",
                          "assign"
                        ]
                      }
                    },
                    "subjects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "subject": {
                            "type": "string",
                            "enum": [
                              "all",
                              "Task",
                              "Profile",
                              "User",
                              "Role",
                              "AbilityRule",
                              "AuditLog"
                            ]
                          },
                          "conditionFields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "readableFields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "writableFields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "identityFields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        },
                        "required": [
                          "subject",
                          "conditionFields",
                          "readableFields",
                          "writableFields",
                          "identityFields"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "operators": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "actorReferences": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "conditionSchemaVersion": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "actions",
                    "subjects",
                    "operators",
                    "actorReferences",
                    "conditionSchemaVersion"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}